+++ src/BDSBeamlineIntegral.cc
@@ -21,6 +21,8 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "BDSComponentFactory.hh"
 #include "BDSDebug.hh"
 #include "BDSException.hh"
+#include "BDSFieldEMRFCavity.hh"
+#include "BDSGlobalConstants.hh"
 #include "BDSParticleDefinition.hh"
 #include "BDSUtilities.hh"
 
@@ -34,14 +36,17 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 
 #include <cmath>
 
-
 BDSBeamlineIntegral::BDSBeamlineIntegral(const BDSParticleDefinition& incomingParticle,
                                          G4double T0In,
-                                         G4double integratedArcLength):
+                                         G4double integratedArcLength,
+                                         G4bool   integrateKineticEnergyIn):
+  integrateKineticEnergy(integrateKineticEnergyIn),
   synchronousTAtEnd(T0In),
   synchronousTAtMiddleOfLastElement(T0In),
   arcLength(integratedArcLength),
-  designParticle(incomingParticle)
+  designParticle(incomingParticle),
+  changeOfEnergyEncountered(false),
+  rigidityCount(0)
 {;}
 
 BDSBeamlineIntegral::~BDSBeamlineIntegral()
@@ -51,40 +56,51 @@ void BDSBeamlineIntegral::Integrate(const GMAD::Element& componentAsDefined)
 {
   // length
   G4double thisComponentArcLength = componentAsDefined.l*CLHEP::m;
+  if (componentAsDefined.type == GMAD::ElementType::_RBEND)
+    {// reuse component factor code to work out angle / field and true arc length
+      G4double componentChordLength;
+      G4double field;
+      G4double angle;
+      BDSComponentFactory::CalculateAngleAndFieldRBend(&componentAsDefined,
+                                                       designParticle.BRho(),
+                                                       thisComponentArcLength,
+                                                       componentChordLength,
+                                                       field,
+                                                       angle);
+    }
   arcLength += thisComponentArcLength;
   
-  G4double v0 = designParticle.Velocity(); // velocity at entrance of element
+  G4double v0 = designParticle.Velocity(); // current velocity at the entrance of element
 
   // calculate change in velocity and kinetic energy
   G4double dEk = 0;
-  
-  G4double particleCharge = designParticle.Charge();
-  G4double phase = componentAsDefined.phase * CLHEP::rad;
-  G4double cosPhase = std::cos(phase);
-  G4double frequency = componentAsDefined.frequency * CLHEP::hertz;
   
   switch (componentAsDefined.type)
     {
     case GMAD::ElementType::_RF:
       {
-        // field model
-        BDSCavityFieldType tp = BDS::DetermineCavityFieldType(componentAsDefined.cavityFieldType);
-        G4double eField = BDSComponentFactory::EFieldFromElement(&componentAsDefined, thisComponentArcLength);
+        G4double particleCharge = designParticle.Charge();
+        G4double phase = componentAsDefined.phase * CLHEP::rad;
+        G4double frequency = componentAsDefined.frequency * CLHEP::hertz;
+        // field model - reproduce behaviour in component factory of default in global constants
+        G4String compCFTN = componentAsDefined.cavityFieldType;
+        G4String cftName = compCFTN.empty() ? BDSGlobalConstants::Instance()->CavityFieldType() : compCFTN;
+        BDSCavityFieldType tp = BDS::DetermineCavityFieldType(cftName);
+        BDSFieldType fieldType = BDS::FieldTypeFromCavityFieldType(tp);
+        G4double eField = BDSComponentFactory::EFieldFromElement(&componentAsDefined, fieldType, thisComponentArcLength, designParticle);
         switch (tp.underlying())
           {
           case BDSCavityFieldType::constantinz:
             {
-              dEk = particleCharge * eField * thisComponentArcLength * cosPhase;
+              dEk = particleCharge * eField * thisComponentArcLength * std::cos(phase);
               break;
             }
           case BDSCavityFieldType::pillbox:
             {
               if (!BDS::IsFinite(frequency)) // protect against zero division
                 {throw BDSException(__METHOD_NAME__, "for a pillbox cavity field, the frequency must be non-zero");}
-              G4double rfWavelength = CLHEP::c_light / frequency;
-              G4double piGOverBetaLambda = CLHEP::pi * thisComponentArcLength / designParticle.Beta() * rfWavelength;
-              G4double transitTimeFactor = std::sin(piGOverBetaLambda) / piGOverBetaLambda;
-              dEk = particleCharge * eField * thisComponentArcLength * transitTimeFactor * cosPhase;
+              G4double transitTimeFactor = BDSFieldEMRFCavity::TransitTimeFactor(frequency, phase, thisComponentArcLength, designParticle.Beta());
+              dEk = particleCharge * eField * thisComponentArcLength * transitTimeFactor;
               break;
             }
           default:
@@ -96,16 +112,39 @@ void BDSBeamlineIntegral::Integrate(const GMAD::Element& componentAsDefined)
       {break;} // no action
     }
   
+  G4bool changeOfEnergyEncounteredThisTime = BDS::IsFinite(dEk, 1e-7);
+  changeOfEnergyEncountered = changeOfEnergyEncountered || changeOfEnergyEncounteredThisTime;
+  // count the number of times rigidity changes so we can uniquely name reused components
+  if (changeOfEnergyEncounteredThisTime)
+    {rigidityCount++;}
+  
   // momentum and therefore BRho
-  designParticle.ApplyChangeInKineticEnergy(dEk);
+  if (integrateKineticEnergy)
+    {designParticle.ApplyChangeInKineticEnergy(dEk);}
 
   G4double v1 = designParticle.Velocity(); // velocity at the end
   G4double vMean = (v1 + v0) / 2.0;
   G4double dT = thisComponentArcLength / vMean;
 
-  G4double dTMiddle = thisComponentArcLength / ( (0.5*(v1-v0) + v0 ) / 2.0);
+  G4double dTMiddle = 0.5*thisComponentArcLength / (0.5*(v1-v0) + v0);
   synchronousTAtMiddleOfLastElement = synchronousTAtEnd + dTMiddle;
   
   // time - now at the start of the next component / end of this component
   synchronousTAtEnd += dT;
+}
+
+G4double BDSBeamlineIntegral::ProvideSynchronousTAtCentreOfNextElement(const GMAD::Element* el) const
+{
+  BDSBeamlineIntegral copy(*this);
+  copy.Integrate(*el);
+  G4double synchronousTAtMiddle = copy.synchronousTAtMiddleOfLastElement;
+  return synchronousTAtMiddle;
+}
+
+G4double BDSBeamlineIntegral::ProvideSynchronousTAtEndOfNextElement(const GMAD::Element* el) const
+{
+  BDSBeamlineIntegral copy(*this);
+  copy.Integrate(*el);
+  G4double synchronousTAtEndLocal = copy.synchronousTAtEnd;
+  return synchronousTAtEndLocal;
 }