+++ src/BDSMagnetStrength.cc
@@ -62,7 +62,6 @@ const std::vector<G4String> BDSMagnetStrength::keys = {
   "kg",              // gabor lens focusing strength
   "plasmaEfield",    // magnitude of electric strength of confined Gabor lens plasma
   "frequency",       // frequency for time varying field (presumably em)
-  "tOffset",         // tOffset resulting in a phase for time varying field
   "phase",           // phase for time varying field
   "synchronousT0",   // global T0 for the synchronous particle at the centre of the object
   "equatorradius",   // radius from axis at which field goes to 0
@@ -141,7 +140,6 @@ const std::map<G4String, BDSMagnetStrength::unitsFactors> BDSMagnetStrength::uni
     {"k11s"          , {"",    1.0}},
     {"k12s"          , {"",    1.0}},
     {"frequency"     , {"",    CLHEP::megahertz}},
-    {"tOffset"       , {"s",   CLHEP::s}},
     {"phase"         , {"rad", CLHEP::rad}},
     {"synchronousT0" , {"s",   CLHEP::s}},
     {"equatorradius" , {"m",   CLHEP::m}},